+++ src/newbank/server/Account.java
@@ -10,19 +10,28 @@ public class Account {
 		this.accountName = accountName;
 		this.balance = openingBalance;
 	}
-	
+
+	@Override
 	public String toString() {
 		return (accountName + ": " + balance);
 	}
 	 public String getAccountName() {
         return accountName;
     }
-	public double getBalance() {
-        return balance;
-    }
+
+
 	public void setBalance(double balance) {
         this.balance = balance;
     }
 	
 
+	public String getAccountByName() {
+
+		return  accountName;
+	}
+
+	public double getBalance(){
+
+		return balance;
+	}
 }