+++ src/newbank/server/Customer.java
@@ -1,27 +1,48 @@
 package newbank.server;
 
+import newbank.server.service.security.PasswordManagerService;
 import java.util.ArrayList;
 
 public class Customer {
 
-	private final String password;
 	private final ArrayList<Account> accounts;
+	private final String passwordHash;
+	private final String passwordSalt;
 
-	public Customer(String password) {
-        this.password = password;
-        accounts = new ArrayList<>();
+	public Customer(String userInputPassword) {
+
+		try {
+			PasswordManagerService.HashResult result = PasswordManagerService.hashPassword(userInputPassword);
+
+			this.passwordHash = result.hash;
+			this.passwordSalt = result.salt;
+
+			//System.out.println("this is the pass stored" + passwordHash + " " + passwordSalt);
+
+		} catch (Exception e) {
+            throw new RuntimeException("Error hashing password process", e);
+        }
+
+        this.accounts = new ArrayList<>();
     }
-	
 
-	public boolean checkPassword(String pw) {
-        return password.equals(pw);
+
+	public boolean checkPassword(String passwordToCheck){
+
+		try {
+			//use the new verify method
+			return PasswordManagerService.verify(passwordToCheck, passwordHash, passwordSalt);
+
+		} catch (Exception e) {
+
+            throw new RuntimeException("Password verification failed", e);
+        }
     }
 
     public String accountsToString() {
         StringBuilder sb = new StringBuilder();
         for (Account a : accounts) {
-            if (sb.length() > 0) {
-            /*if (!sb.isEmpty()) {*/ /*removed this part */
+            if (!sb.isEmpty()) {
                 sb.append("\n");
             }
             sb.append("> ").append(a.toString());
@@ -32,6 +53,15 @@ public class Customer {
 
 	public void addAccount(Account account) {
 		accounts.add(account);		
+	}
+
+	public Account getAccountByName(String name) {
+		for (Account a : accounts) {
+			if (a.getAccountByName().equalsIgnoreCase(name)) {
+				return a;
+			}
+		}
+		return null;
 	}
 
 public boolean hasAccount(String accountName) {